a trigger's a wonderful thing
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonData/trunk@45382 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Load/lib/sql/comments/searchableCommentText.sql
@@ -69,24 +69,27 @@ on apidb.TextSearchableComment(content)
 indextype is ctxsys.context
 parameters('DATASTORE CTXSYS.DEFAULT_DATASTORE SYNC (ON COMMIT)');
 
-create or replace trigger comments2.comments_insert
+create or replace trigger comments_insert
   after insert on comments2.comments
   for each row
 declare
   userinfo varchar2(1000);
 begin
-  select first_name || ' ' || last_name || '(' || organization || ')'
-  into userinfo
-  from userlogins3.users
-  where user_id = :new.user_id;
 
-  insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
-  values (:new.comment_id, :new.stable_id, :new.project_name, :new.organism,
-          :new.headline || '|' || :new.content || '|' || userinfo || author_list(:new.comment_id));
+  if :new.comment_target_id = 'gene' then
+    select first_name || ' ' || last_name || '(' || organization || ')'
+    into userinfo
+    from userlogins3.users
+    where user_id = :new.user_id;
+
+    insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
+    values (:new.comment_id, :new.stable_id, :new.project_name, :new.organism,
+            :new.headline || '|' || :new.content || '|' || userinfo || apidb.author_list(:new.comment_id));
+  end if;
 end;
 /
 
-create or replace trigger comments2.comments_delete
+create or replace trigger comments_delete
   after delete on comments2.comments
   for each row
 begin
@@ -95,7 +98,7 @@ begin
 end;
 /
 
-create or replace trigger comments2.comments_update
+create or replace trigger comments_update
   after update on comments2.comments
   for each row
 declare
@@ -105,51 +108,77 @@ begin
   delete from apidb.TextSearchableComment
   where comment_id = :old.comment_id;
 
-  select first_name || ' ' || last_name || '(' || organization || ')'
-  into userinfo
-  from userlogins3.users
-  where user_id = :new.user_id;
+  if :new.comment_target_id = 'gene' then
+    begin
+      select first_name || ' ' || last_name || '(' || organization || ')'
+      into userinfo
+      from userlogins3.users
+      where user_id = :new.user_id;
+    exception
+      when NO_DATA_FOUND then
+        userinfo := ' ()'; -- literals from userinfo string
+    end;
 
-  select apidb.author_list(:new.comment_id)
-  into authorinfo
-  from dual;
+    select apidb.author_list(:new.comment_id)
+    into authorinfo
+    from dual;
 
-  insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
-  values (:new.comment_id, :new.stable_id, :new.project_name, :new.organism,
-          :new.headline || '|' || :new.content || '|' || userinfo);
+    insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
+    values (:new.comment_id, :new.stable_id, :new.project_name, :new.organism,
+            :new.headline || '|' || :new.content || '|' || userinfo || apidb.author_list(:new.comment_id));
+
+    insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
+    select :new.comment_id, stable_id, :new.project_name, :new.organism,
+            :new.headline || '|' || :new.content || '|' || userinfo || apidb.author_list(:new.comment_id)
+    from comments2.CommentStableId
+    where comment_id = :new.comment_id
+      and stable_id != :new.stable_id;
+  end if;
 end;
 /
 
-create or replace trigger comments2.csi_insert
+create or replace trigger csi_insert
   after insert on comments2.commentStableId
   for each row
 declare
   userinfo varchar2(1000);
 begin
-  select first_name || ' ' || last_name || '(' || organization || ')'
-  into userinfo
-  from userlogins3.users
-  where user_id = (select user_id from comments2.comments where comment_id = :new.comment_id);
+    begin
+      select first_name || ' ' || last_name || '(' || organization || ')'
+      into userinfo
+      from userlogins3.users
+      where user_id = (select user_id from comments2.comments where comment_id = :new.comment_id);
+    exception
+      when NO_DATA_FOUND then
+        userinfo := ' ()'; -- literals from userinfo string
+    end;
 
   insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
   select comment_id, :new.stable_id, project_name, organism,
           headline || '|' || content || '|' || userinfo || apidb.author_list(comment_id)
   from comments2.comments
-  where comment_id = :new.comment_id;
+  where comment_id = :new.comment_id
+    and comment_target_id = 'gene'
+    and stable_id != :new.stable_id; -- don't duplicate comment-gene pairs
 end;
 /
 
-create or replace trigger comments2.csi_delete
+create or replace trigger csi_delete
   after delete on comments2.CommentStableId
   for each row
 begin
   delete from apidb.TextSearchableComment
   where comment_id = :old.comment_id
-    and source_id = :old.stable_id;
+    and source_id = :old.stable_id
+    -- last condition ensures that the mapping being deleted
+    -- is not duplicated in the comment table itself
+    and source_id != (select stable_id
+                      from comments2.comments
+                      where comment_id = :old.comment_id);
 end;
 /
 
-create or replace trigger comments2.csi_update
+create or replace trigger csi_update
   after update on comments2.commentStableId
   for each row
 declare
@@ -157,47 +186,116 @@ declare
 begin
   delete from apidb.TextSearchableComment
   where comment_id = :old.comment_id
-    and source_id = :old.stable_id;
+    and source_id = :old.stable_id
+    and (:old.comment_id, :old.stable_id)
+        not in (select comment_id, stable_id from comments2.Comments);
 
-  select first_name || ' ' || last_name || '(' || organization || ')'
-  into userinfo
-  from userlogins3.users
-  where user_id = (select user_id from comments2.comments where comment_id = :new.comment_id);
+    begin
+      select first_name || ' ' || last_name || '(' || organization || ')'
+      into userinfo
+      from userlogins3.users
+      where user_id = (select user_id from comments2.comments where comment_id = :new.comment_id);
+    exception
+      when NO_DATA_FOUND then
+        userinfo := ' ()'; -- literals from userinfo string
+    end;
 
   insert into apidb.TextSearchableComment (comment_id, source_id, project_id, organism, content)
   select comment_id, :new.stable_id, project_name, organism,
           headline || '|' || content || '|' || userinfo || apidb.author_list(comment_id)
   from comments2.comments
-  where comment_id = :new.comment_id;
+  where comment_id = :new.comment_id
+    and comment_target_id = 'gene';
+
 end;
 /
 
-create or replace trigger comments2.author_trigger
-  after insert or update or delete on comments2.CommentReference
-  for each row
+-- Changes to CommentReference can trigger only updates to TextSearchableComment.
+-- But the recording of affected comment IDs must be separated from the updating
+-- of those comments' content strings, to avoid the dread
+-- "ORA-04091: table CommentReference is mutating, trigger/function may not see it".
+-- This is a issue for CommentReference (unlike Comments and CommentStableId) because
+-- the affected
+-- 
+-- hence, three triggers, plus a package to let them share info
+
+create or replace package cmntRef_trggr_pkg
+as
+    type commentIdList is table of number index by binary_integer;
+         stale    commentIdList;
+         empty    commentIdList;
+     end;
+/
+
+-- a statement-level trigger runs before a row-level trigger when both are fired
+-- by a single SQL statement
+-- http://docs.oracle.com/cd/B19306_01/server.102/b14220/triggers.htm#sthref3278
+
+-- once per statement, initialize the list of comments IDs affected by the triggering change to CommentReference.
+create or replace trigger cmntRef_setup
+before insert or update or delete on comments2.CommentReference
+begin
+    cmntRef_trggr_pkg.stale := cmntRef_trggr_pkg.empty;
+end;
+/
+
+-- once per row, note the ID of the comment that had an author inserted
+create or replace trigger cmntRef_markInsertedId
+before insert on comments2.CommentReference
+for each row
+declare
+    i    number default cmntRef_trggr_pkg.stale.count+1;
+begin
+    cmntRef_trggr_pkg.stale(i) := :new.comment_id;
+end;
+/
+
+-- once per row, note the ID of the comment that had an author deleted
+create or replace trigger cmntRef_markDeletedId
+before delete on comments2.CommentReference
+for each row
+declare
+    i    number default cmntRef_trggr_pkg.stale.count+1;
+begin
+    cmntRef_trggr_pkg.stale(i) := :old.comment_id;
+end;
+/
+
+-- once per row, note the ID of the comment that had an author updated
+create or replace trigger cmntRef_markUpdatedId
+before update on comments2.CommentReference
+for each row
+declare
+    i    number default cmntRef_trggr_pkg.stale.count+1;
+begin
+    cmntRef_trggr_pkg.stale(i) := :old.comment_id;
+    cmntRef_trggr_pkg.stale(i+1) := :new.comment_id;
+end;
+/
+
+-- after the statement, update the content of affected TextSearchableComment records
+create or replace trigger cmntRef_updateTsc
+   after insert or update or delete on comments2.CommentReference
 declare
   userinfo varchar2(1000);
 begin
-  select first_name || ' ' || last_name || '(' || organization || ')'
-  into userinfo
-  from userlogins3.users
-  where user_id = (select user_id from comments2.comments where comment_id = :old.comment_id);
+    for i in 1 .. cmntRef_trggr_pkg.stale.count loop
 
-  update apidb.TextSearchableComment
-  set content = (select headline || '|' || content || '|' || userinfo || apidb.author_list(comment_id)
-                 from comments2.comments
-                 where comment_id = :old.comment_id)
-  where comment_id = :old.comment_id;
+        begin
+          select first_name || ' ' || last_name || '(' || organization || ')'
+          into userinfo
+          from userlogins3.users
+          where user_id = (select user_id from comments2.comments where comment_id = cmntRef_trggr_pkg.stale(i));
+        exception
+          when NO_DATA_FOUND then
+            userinfo := ' ()'; -- literals from userinfo string
+        end;
 
-  select first_name || ' ' || last_name || '(' || organization || ')'
-  into userinfo
-  from userlogins3.users
-  where user_id = (select user_id from comments2.comments where comment_id = :new.comment_id);
-
-  update apidb.TextSearchableComment
-  set content = (select headline || '|' || content || '|' || userinfo || apidb.author_list(comment_id)
-                 from comments2.comments
-                 where comment_id = :new.comment_id)
-  where comment_id = :new.comment_id;
+        update apidb.TextSearchableComment
+        set content = (select headline || '|' || content || '|' || userinfo || apidb.author_list(comment_id)
+                       from comments2.comments
+                       where comment_id = cmntRef_trggr_pkg.stale(i))
+        where comment_id = cmntRef_trggr_pkg.stale(i);
+    end loop;
 end;
 /
